Notify user that there is a SES7.1 upgrade available.
Let the Dashboard module configure the MOTD itself instead doing that via cephadm or ceph-salt. The new approach seems to be the easier way.
To do so, the Dashboard plugin 'ses7p_motd' has been introduced. This should prevent or reduce the number of merge conflicts.
The plugin makes sure it will not configure the MOTD on every start of the Ceph Dashboard module. This way it is possible that the administrator can disable the MOTD manually.
The MOTD will use the 'danger' severity, thus the message can not be closed by the user and it will always appear when the user signs in to the Dashboard UI.

Fixes: https://jira.suse.com/browse/SES-2697

Signed-off-by: Volker Theile <vtheile@suse.com>